% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam_scores}
\alias{exam_scores}
\title{Production and farm value of maple products in Canada}
\format{
A data frame with columns:
\describe{
 \item{score}{A value between 0 and 100}
 \item{hrs_sleep_per_week}{Hours slept per week.}
 \item{motivation}{Student motivation. Likert scale from 1 to 10.}
 \item{note_mode}{The method of note-taking. 1 = Paper, 2 = Computer.}
}
}
\source{
Generated based on published results from various studies.
}
\usage{
exam_scores
}
\description{
Production and farm value of maple products in Canada
}
\examples{
\dontrun{
 exam_scores
}
}
\keyword{datasets}
