% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam_scores}
\alias{exam_scores}
\title{Exam scores and sleep}
\format{
A data frame with columns:
\describe{
 \item{exam1}{Score on the first exam. A value between 0 and 100.}
 \item{motivation}{Self-reported motivation score. Likert scale from 1 to 10.}
 \item{note_mode}{The method of note-taking. 1 = Paper, 2 = Computer.}
 \item{exam2}{Score on the second exam. A value between 0 and 100.}
 \item{anxiety}{Self-reported level of anxiety. 1 = low, 2 = medium, 3 = high.}
 \item{sleep_week}{Minutes of sleep in the week before exam 1.}
 \item{sleep_day}{Minutes of sleep in the day before exam 1.}
}
}
\source{
An artifical dataset created for demonstrative purposes only. Generated based on published results from various studies.
}
\usage{
exam_scores
}
\description{
Exam scores and sleep
}
\examples{
\dontrun{
 exam_scores
}
}
\keyword{datasets}
