% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_tutorial.R
\name{launch_tutorial}
\alias{launch_tutorial}
\title{Launch Tutorials}
\usage{
launch_tutorial(tutorial)
}
\arguments{
\item{tutorial}{character, representing the name of a tutorial to run. Options include:
\enumerate{
\item \code{'intro-to-r'}: A basic introduction to how to use R.
\item \code{'data-frames'}: An introduction to using data frames in R.
\item \code{'categorical-descriptives'}: Using R to do basic descriptive statistics with categorical data.
\item \code{'numerical-descriptives'}: Using R to do basic descriptive statistics with numerical data.
\item \code{'regression'}: Ordinary least squares regression using R.
\item \code{'t-tests'}: An introduction to doing t-tests in R.
\item \code{'anova'}: Using R for ANOVA.
}}
}
\value{
None.
}
\description{
This function is used to launch tutorials.
}
\examples{
launch_tutorial('intro-to-r')
}
