% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_tutorial.R
\name{launch_tutorial}
\alias{launch_tutorial}
\title{Launch Tutorials}
\usage{
launch_tutorial(tutorial)
}
\arguments{
\item{tutorial}{character, representing the name of a tutorial to run. Options include:
1) 'intro-to-r': A basic introduction to how to use R.}
}
\value{
None.
}
\description{
This function is used to launch tutorials.
}
\examples{
launch_tutorial('intro-to-r')
}
